# PoC: Comparing aiohttp to requests