it seems like running aiohttp is naively much faster than requests within a threadpool with an equivalent number of workers. I ran 1000 requests

concurrency = 1, requests with ThreadPool took 212.26 seconds, aiohttp took 49.51seconds :exploding_head:
concurrency = 5, requests with ThreadPool took 41.72 seconds, aiohttp took 12.21 seconds
concurrency = 10, requests with ThreadPool took 21.49 seconds, aiohttp took 6.18 seconds

The test script was: 
```python
import time
import requests
from concurrent.futures import ThreadPoolExecutor
import aiohttp
import asyncio

URL = "https://api.stripe.com/v1/customers"  # Replace with the URL you want to test
API_KEY="<>"
h={"Authorization": "Basic " + f"{API_KEY}:".encode("utf-8").hex()}
qparams = {"limit": 100}
NUM_REQUESTS = 1000
CONCURRENCY = 5 

# Using requests with ThreadPool
def fetch_with_requests(url):
    try:
        response = requests.get(url, headers=h, params=qparams)
        return response.status_code
    except requests.RequestException as e:
        print(e)
        return None

def run_requests_threadpool():
    with ThreadPoolExecutor(max_workers=CONCURRENCY) as executor:
        futures = [executor.submit(fetch_with_requests, URL) for _ in range(NUM_REQUESTS)]
        return [f.result() for f in futures]

# Using aiohttp
async def fetch_with_aiohttp(session, url, semaphore: asyncio.Semaphore):
    async with semaphore:
        async with session.get(url, headers=h, params=qparams) as response:
            return response.status

async def run_aiohttp():
    semaphore = asyncio.Semaphore(CONCURRENCY)
    async with aiohttp.ClientSession() as session:
        tasks = [fetch_with_aiohttp(session, URL, semaphore) for _ in range(NUM_REQUESTS)]
        return await asyncio.gather(*tasks)

def main():
    # Test requests with ThreadPool
    start_time = time.time()
    run_requests_threadpool()
    requests_time = time.time() - start_time
    print(f"requests with ThreadPool took {requests_time:.2f} seconds")

    # Test aiohttp
    start_time = time.time()
    asyncio.run(run_aiohttp())
    aiohttp_time = time.time() - start_time
    print(f"aiohttp took {aiohttp_time:.2f} seconds")

if __name__ == "__main__":
    main()
```

I wondered if this meant we can get a 4x improvement for free by transparently swapping out aiohttp for requests under the hood. It doesn’t seem like it. 

For the same stripe dataset (pulling data for 1300 customers) a pure requests approach took 11seconds and the sequential aiohttp approach took 14seconds

I suspected this might be because I’m spinning up one event loop per request

The profiling didn’t give a clear explanation

Profiling of the `requests` based approach within the CDK
```
(.venv) ➜  source-stripe git:(sherif/aiohttp-to-requests-test) ✗ awk '/^[ ]*[0-9]/ { print $0 }' async.perf | sort -k2,2n | tail -n 30

      601    0.015    0.000    0.015    0.000 {built-in method io.open_code}
    65150    0.015    0.000    0.026    0.000 validators.py:76(type_check)
     6666    0.016    0.000    0.025    0.000 parse.py:437(urlsplit)
    65201    0.017    0.000    0.077    0.000 _pmap.py:168(__getitem__)
    65201    0.017    0.000    0.148    0.000 validators.py:355(is_type)
    87367    0.017    0.000    0.136    0.000 _validators.py:272(<genexpr>)
      739    0.018    0.000   12.215    0.017 selectors.py:554(select)
    32575    0.018    0.000    0.065    0.000 transform.py:75(__normalize)
    65/40    0.018    0.000    0.064    0.002 {built-in method _imp.exec_dynamic}
7825/1305    0.019    0.000    0.325    0.000 _validators.py:276(properties)
     1304    0.023    0.000   12.134    0.009 abstract_source.py:301(_read_full_refresh)
     1316    0.023    0.000    0.030    0.000 encoder.py:204(iterencode)
    65202    0.024    0.000    0.060    0.000 _pmap.py:158(_getitem)
1544/1505    0.024    0.000    0.068    0.000 {built-in method builtins.__build_class__}
    32575    0.025    0.000    0.035    0.000 transform.py:89(default_convert)
    65205    0.026    0.000    0.036    0.000 _pmap.py:152(_get_bucket)
    65201    0.028    0.000    0.131    0.000 _types.py:66(is_type)
330286/330248    0.028    0.000    0.032    0.000 {built-in method builtins.isinstance}
      594    0.030    0.000    0.030    0.000 {built-in method marshal.loads}
    33911    0.031    0.000    0.185    0.000 _validators.py:269(type)
      447    0.035    0.000    0.035    0.000 {method 'recv' of '_socket.socket' objects}
       38    0.042    0.001    0.043    0.001 decoder.py:343(raw_decode)
33944/1305    0.046    0.000    0.424    0.000 validators.py:296(iter_errors)
44302/2606    0.046    0.000    0.419    0.000 transform.py:136(normalizator)
    65/62    0.053    0.001    0.055    0.001 {built-in method _imp.create_dynamic}
     1306    0.064    0.000    0.104    0.000 entrypoint.py:148(airbyte_message_to_string)
     1303    0.082    0.000    0.593    0.000 record_helper.py:14(stream_data_to_airbyte_message)
      796    0.083    0.000    0.083    0.000 {method 'read' of '_io.BufferedReader' objects}
     1321    0.214    0.000    0.214    0.000 {built-in method builtins.print}
      814   12.195    0.015   12.195    0.015 {method 'control' of 'select.kqueue' objects}
```
Profiling of the synchronous approach
```
(.venv) ➜  source-stripe git:(sherif/aiohttp-to-requests-test) ✗ awk '/^[ ]*[0-9]/ { print $0 }' sync.perf | sort -k2,2n | tail -n 30

      601    0.013    0.000    0.013    0.000 {built-in method io.open_code}
     6801    0.013    0.000    0.022    0.000 parse.py:437(urlsplit)
      423    0.015    0.000    0.015    0.000 {built-in method builtins.compile}
    65150    0.015    0.000    0.025    0.000 validators.py:76(type_check)
    65201    0.017    0.000    0.074    0.000 _pmap.py:168(__getitem__)
    65201    0.017    0.000    0.144    0.000 validators.py:355(is_type)
    87367    0.017    0.000    0.135    0.000 _validators.py:272(<genexpr>)
    32575    0.018    0.000    0.062    0.000 transform.py:75(__normalize)
        1    0.022    0.022    0.022    0.022 {built-in method _socket.getaddrinfo}
       38    0.022    0.001    0.022    0.001 decoder.py:343(raw_decode)
    65202    0.022    0.000    0.058    0.000 _pmap.py:158(_getitem)
1541/1502    0.023    0.000    0.065    0.000 {built-in method builtins.__build_class__}
     1316    0.024    0.000    0.030    0.000 encoder.py:204(iterencode)
    32575    0.024    0.000    0.033    0.000 transform.py:89(default_convert)
      592    0.025    0.000    0.025    0.000 {built-in method marshal.loads}
    65205    0.026    0.000    0.035    0.000 _pmap.py:152(_get_bucket)
334272/334264    0.026    0.000    0.030    0.000 {built-in method builtins.isinstance}
    33911    0.027    0.000    0.179    0.000 _validators.py:269(type)
    65201    0.027    0.000    0.127    0.000 _types.py:66(is_type)
        1    0.035    0.035    0.035    0.035 {method 'do_handshake' of '_ssl._SSLSocket' objects}
7825/1305    0.040    0.000    0.338    0.000 _validators.py:276(properties)
    65/62    0.041    0.001    0.044    0.001 {built-in method _imp.create_dynamic}
44302/2606    0.043    0.000    0.426    0.000 transform.py:136(normalizator)
        1    0.045    0.045    0.045    0.045 {method 'connect' of '_socket.socket' objects}
33944/1305    0.047    0.000    0.431    0.000 validators.py:296(iter_errors)
     1306    0.063    0.000    0.104    0.000 entrypoint.py:148(airbyte_message_to_string)
      796    0.079    0.000    0.079    0.000 {method 'read' of '_io.BufferedReader' objects}
     1303    0.079    0.000    0.592    0.000 record_helper.py:14(stream_data_to_airbyte_message)
     1306    0.181    0.000    0.181    0.000 {built-in method builtins.print}
      141   11.436    0.081   11.436    0.081 {method 'read' of '_ssl._SSLSocket' objects}
```